This release includes bug fixes,
including:

Fixes an issue that cause the application to crash on startup. (It was being caused by the php code not being able to connect with the application. Instead of http://localhost it required the use of http://10.0.2.2, as the android emulator refers to the local pc with that address.)

Fixed DatePicker & TimePicker modules, added Listeners to extract user's input selection.

New features:

Added permissions in AndroidManifest.xml to allow the app to connect to the internet & check the network status.

Added Java SE 7 support via build.gradle to implement higher level API.

Added AsyncTask class to implement a network connection while running in a background thread, thus not freezing the main thread.

Added MySql Database Connection via php code (refer to the OnCampusResourceManagementPHP repository for the php code)

Added HttpManager class to set up the connection to the php file and import its output.

Added a JSONParser class to parse the json encoded output generated by the php code.

Added a PlainOldJavaObjects class to store and retrieve data after it has been parsed by the JSONParser class.